% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\docType{data}
\name{complexity_dat}
\alias{complexity_dat}
\title{Artificial data set to illustrate the calculation of the complexity frontier.
A data set illustrating the form and style how the data should be structured to be used by the function \emph{complexity_frontier}.}
\format{
## `complexity_dat`
A data frame with 995 rows and 7 columns:
\describe{
  \item{geo}{The ids of the spatial units, e.g. regions.}
  \item{year}{The year for which the calculation is done.}
  \item{appln_id}{Unique patent document identifier.}
  \item{tech}{Technology identifier with less digits than patent class.}
  \item{structural}{Random number between 1 and 12 representing the value of structural complexity.}
}
}
\source{
None
}
\usage{
complexity_dat
}
\description{
Artificial data set to illustrate the calculation of the complexity frontier.
A data set illustrating the form and style how the data should be structured to be used by the function \emph{complexity_frontier}.
}
\keyword{datasets}
