% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{complexity_estimation}
\alias{complexity_estimation}
\title{Complexity Estimations}
\usage{
complexity_estimation(
  patdat,
  pats_window,
  mw = 3,
  node.sample = 125,
  reps = 200
)
}
\arguments{
\item{reps}{}
}
\description{
Complexity Estimations
}
