% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{giant.component}
\alias{giant.component}
\title{Giant component}
\usage{
giant.component(graph)
}
\arguments{
\item{graph}{An igraph object from which the giant component needs to be extracted.}
}
\value{
The function returns an igraph object representing the giant component of a network
}
\description{
Helper function for the calculation of the network diversity score of \insertCite{Emmert-Streib2012;textual}{GeoInno} and the structural diversity complexity measure of \insertCite{Broekel2019;textual}{GeoInno}.
}
\examples{
my.graph <- igraph::random.graph.game(p.or.m = 1/10, n=10)
giant.component(my.graph)
}
