% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{nds}
\alias{nds}
\title{Function nds}
\usage{
nds(g, node.sample, reps, ...)
}
\arguments{
\item{g}{An igraph network main component.}

\item{node.sample}{The sample of nodes for which the calculation is to be done.}

\item{reps}{The number of repetition of the bootstrap.}
}
\value{
A numeric value of the NDS-score, after a log-transformation and multiplication with -1.
}
\description{
Background function for the estimation of the structural diversity measure by \insertCite{Emmert-Streib2012;textual}{GeoInno}. It corresponds to the calculation of the network diversity score (NDS) of \insertCite{Broekel2019;textual}{GeoInno} for a binary network. The network must be binary and connected (giant component).
}
