% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{complexity_frontier}
\alias{complexity_frontier}
\title{Complexity frontier}
\usage{
complexity_frontier(complex.dat, percentil = 5)
}
\value{
The function returns a data.frame with the name complexity.frontier including the following information:
\itemize{
  \item{\emph{tech}}{ focal technology}
  \item{\emph{year}}{ year for which the calculation is done}#' }
}
\description{
The function ...
}
\examples{
complexity_frontier(complex.dat)
}
\references{
\insertAllCited{}
}
