% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{create_sample_data}
\alias{create_sample_data}
\title{Create sample data}
\usage{
create_sample_data(num.pat = 200, geos = 10)
}
\arguments{
\item{num.pat}{Number of patents.}

\item{geos}{Number of spatial units.}
}
\value{
The function returns a data.frame suited to be used with the function \emph{structural_complexity}.
\itemize{
  \item{geo}{The ids of the spatial unit, e.g. regions.}
  \item{year}{The year for which the calculation is done.}
  \item{appln_id}{Unique patent document identifier.}
  \item{cpc}{Patent class identifier with more digits than technology.}
  \item{tech}{Technology identifier with less digits than patent class.}
  \item{structural}{Random number between 1 and 12 representing the value of structural complexity.}
  }
}
\description{
Create sample data
}
\examples{
create_sample_data(num.pat=200, geos=10)
}
