% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_calculation.R
\name{create_sample_data}
\alias{create_sample_data}
\title{Create sample data}
\usage{
create_sample_data(num.pat = 200)
}
\arguments{
\item{num.pat}{Number of patents}
}
\value{
The function returns a data.frame suited to be used with other functions in the package
}
\description{
Create sample data
}
\examples{
create_sample_data(200)
}
